implement utf8, part 16
