+++ src/openrct2.c
@@ -195,6 +195,7 @@ bool openrct2_initialise()
 	addhook(0x006BB76E, (int)sound_play_panned, 0, (int[]){EAX, EBX, ECX, EDX, EBP, END}, EAX, 0);	// remove when all callers are decompiled
 	addhook(0x006C42D9, (int)scrolling_text_setup, 0, (int[]){EAX, ECX, EBP, END}, 0, EBX);			// remove when all callers are decompiled
 	addhook(0x006C2321, (int)gfx_get_string_width, 0, (int[]){ESI, END}, 0, ECX);					// remove when all callers are decompiled
+	addhook(0x006C2321, (int)format_string, 0, (int[]){EDI, EAX, ECX, END}, 0, 0);					// remove when all callers are decompiled
 
 	if (!rct2_init())
 		return false;
